Implement an ingress
This uses nginx-ingress, because it has support for forwarding TCP
ports that the GCE ingress does not.  Also, it updates immediately
on change, where the GCE ingress can take upwards of 5 minutes.

The disadvantage is that the nginx-ingress chart provides no way to
include Helm variables in the name of a service targetted by a TCP port.
So the "dev" name is hard-coded.